refactor: refactor bulk creation page
