+++ src/modules/product-bulk-creation/components/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./ProductBulkCreationSection";
