+++ src/modules/product-bulk-creation/containers/ProductBulkCreationContainer.tsx
@@ -1,0 +1,11 @@
+import ProductBulkCreationSection from "../components";
+
+const ProductBulkCreationContainer = () => {
+  return (
+    <article className="flex flex-col items-center justify-center gap-3">
+      <ProductBulkCreationSection />
+    </article>
+  );
+};
+
+export default ProductBulkCreationContainer;
