+++ src/modules/product-bulk-creation/containers/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./ProductBulkCreationContainer";
